+++ sources/gookto_io/InputPax.cpp
@@ -1,12 +1,7 @@
 #include <gookto_io/InputPax.hpp>
 
-#include <gookto_io/tinyapi/tinyxml.h>
-#include <gookto_io/tinyapi/tinystr.h>
 
-#include <string>
-#include <iostream>
-
-InputPax::InputPax(int id, int type, float dpt_time) :
+InputPax::InputPax(int id, int type, double dpt_time) :
 id(id), type(type), dpt_time(dpt_time)
 {};
